Automated g4 rollback.
*** Reason for rollback ***

breaks contacts

*** Original change description ***

When we can't find a forward-declared type,
that type should resolve to an empty type.
This ensures that if someone tries to pass a param or return
value to it, they get a type warning.

DELTA=461  (37 added, 385 deleted, 39 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=302


git-svn-id: https://closure-compiler.googlecode.com/svn/trunk@703 b0f006be-c8cd-11de-a2e8-8d36a3108c74