Automated g4 rollback.
*** Reason for rollback ***

Rollback of a rollback

*** Original change description ***

Automated g4 rollback.

*** Reason for rollback ***

Breaking someone's build

*** Original change description ***

An attempt to describe policies for some of the challenging aspects
of an externs file for the Chrome Extension APIs. Builds on the existing
section describing when to add a pacakge to this externs file to
have:
1. When to add packages to this file
2. Optional parameters
3. Pseduo-types
4. Events
5. Nullability

... description truncated by g4 rollback ...
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=42696808