Change RescopeGlobalSymbols to not rewrite variables that do not cross JSModules.
Reduces code size and speeds up code.

The change tries to leave as much JS code intact as possible if i.e. none of the vars in a var statement have to rewritten, the code is left untouched.

As a followup it may be interesting to the rename-passes again after this pass and/or wrap each module in an immediately called function so that names can be reused. Both may show some benefits after GZIP.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=46751454