Added externs.map file (replaces hard-coded list)
A new file (externs.map) has been added to the externs directory, it
contains the ordered list of default externs that was previously
hard-coded in the CommandLineRunner class.

At runtime the externs.map file is read and the ordered List<String> of
externs is generated.

The externs.map file is not included in the externs.zip file, it is
packed into the compiler.jar file at the same location as the
externs.zip file.