Warn for initializers in for/in loops in all cases, rather than just let and const.
While 'for (var x = 'foo' in bar) {}' is technically valid in ES5, it doesn't really make much sense, and is removed in ES6. So we make it a warning in ES5 mode and an error in ES6 mode.

See discussion at http://esdiscuss.org/topic/initializer-expression-on-for-in-syntax-subject
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=74178222