CheckAccessControlsTest: add 3 tests to document current behavior.
Currently, Closure Compiler issues a warning when a name or property ends in a trailing underscore but is declared to be something other than @private. There are many tests exercising this behavior for properties, but (surprisingly) none for names. This CL adds a few simple cases for names with trailing underscores.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=74250231