JS conformance: add only_apply_to and only_apply_to_regexp fields to conformance.proto.
Currently, the JS conformance framework has no good way of specifying a subset of the compilation unit that should be checked for a particular rule. Of course, this can be achieved with a large whitelist enumerating the subset of the compilation unit that should *not* be checked for the rule. But this is unacceptable for conformance configurations in large projects, because the whitelists would be updated much too often.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=80216762