Automated g4 rollback of changelist 80781799.
*** Reason for rollback ***

This change starts an awful chain of consequences that leads to the current type system's distinction of declared and inferred record types.As is this causes some inferred record types to become declared record types.  It is unclear if it is time well spent to try to fix this. An alternate solution for goog.module is proposed.  However, that solution doesn't help ES6 modules or other module systems that naturally have a lot of type declaration aliases.  Native understanding of ES6 modules is, of course, the correct solution there.

*** Original change description ***

Improve type inferrence for goog.module exports (part 1 of 3):
Fix declaration of type aliases with inferred constants.

***
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=81344430