In RecordTypeBuilder, use an ImmutableSortedMap for record types, instead of an unmodifiable view of a HashMap.
Because the ImmutableSortedMap builder does not accept duplicate keys, this change requires us to move the check for duplicate keys in record types from the JSTypeRegistry to the JSDocInfoParser. This is a nice cleanup anyway, as we don't need to reimplement the check in the new type inference, and it makes more sense for the error to be caught by the parser, than by the internals of the type system.

-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=83794602