Allow JsDoc on export nodes, and copy it to the corresponding var/let/const/function node when transpiling.
Also, add a new Truth Subject for testing Nodes. This allows for a nice way of checking that a Node is a particular type, with a nice error message. For example:

assertEquals(Token.EXPORT, node.getType())

fails with

expected:<169> but was:<132>

but

assertAbout(NODE).that(node).hasType(Token.EXPORT);

fails with

Node is of type SCRIPT not of type EXPORT

Fixes #781 on Github.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=84674107