Minimal patch to allow custom CodeGenerator subclasses.
The built-in behavior of the compiler is to produce only the declared language_out source code. It does this with the primary CodeGenerator, as well as a subtype, TypedCodeGenerator, which additionally emits the declared types in JSDocs.

I would like to have my own subclass of CodeGenerator, to experiment with new language features for AtScript. In particular, I'm overriding the behavior of printing classes to emit an "interface" keyword rather than /** @interface */ class{}. I cannot do this in a CompilerPass because the syntax I'm generating is not valid ES6_TYPED and so it cannot be represented by the AST.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=87382574