Automated g4 rollback of changelist 85777585.
*** Reason for rollback ***

Decrease the number of methods in TypeI and friends. We've recently decided to go about the migration of passes to NTI in a slightly different way.

*** Original change description ***

[NEW TYPE INFERENCE] Convert CodingConvention and its subclasses to use TypeI.

The only classes from rhino.jstype left in coding conventions are JSTypeNative (which we are keeping), StaticScope and StaticSourceFile. We will eventually remove the last two also.

This change requires supporting some of TypeI in the new type inference.

***
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=88341772