PolymerPass: Avoid renaming altogether for generated _set* functions by adding an interface to externs for the element which has the proper _set* functions.
NOTE: I have to add a suppression here for the constructor property in the object literal now because it's not a real constructor and shouldn't be considered as overriding the JS Object "constructor" property. Discussions are ongoing with the Polymer team  to change the name of this property to avoid these collisions.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=91930114