Fix typo in eval call in ClosureBundler.
The code currently does:

(0,eval("..."));

which is a "direct eval call" (See http://ecma262-5.com/ELS5_HTML.htm#Section_10.4.2
for details). I believe the intention was do an indirect call to get a global eval:

(0,eval)("...");

Since we weren't using a global eval before, it would be safer to just remove
the "0," to maintain the same behavior, but I believe doing the indirect call
is the prefered solution, since it more directly simulates the list of <script>
tags.

For example if a function wrapper was added or a "use strict" was added to the
outer script, this wold cause these direct eval calls to have separate
execution contexts, and functions/variables could not be shared (in fact this
is probably why change cr/93455279 was done, the "use strict" broken because of
this eval bug).
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=103319411