Remove UMD pattern when processing CommonJS.
Detects UMD pattern when processing JavaScript modules and rewrites it
to make sure the CommonJS "module.exports" or "exports" statements can
be reached. Checks the condition of an if-statement for an occurrence of
either "module.exports"/"exports" for CommonJS or "define.amd" for
AMD. For an if-statement that is checking for CommonJS, we pull out the
then-branch and remove the rest. For an if-statement that is checking
for AMD, we pull out the else-branch (since it could be checking for
CommonJS) and remove the rest.

For example:

var foobar = {
    foo: "bar"
};

if (typeof define === "function" && define.amd) {
    define([], function() {
        return foobar;
    });
} else if (typeof exports === "object") {
    if (true) {
        module.exports = foobar;
    }
} else {
    this.foobar = foobar;
}

will be translated to:

var module$umd_standard = {},
    foobar$$module$umd_standard = {foo:"bar"},
    module$umd_standard = foobar$$module$umd_standard;