[NTI] Rework the finalization of namespaces, to allow adding class properties to nominal types after namespace finalization.
ProcessScope looks at scopes outer-to-inner, and we can't change this order without major changes (if at all). So, it's very common to see a constructor as the RHS of a @const assignment, without having seen all its class properties yet.

The reason we finalize a namespace is to convert it to a JSType. But when we compute the namespace type for the constructor, we care about the static properties (and the prototype properties), not the class properties (because the class properties are on the instances, not on the constructor). The class properties always live on RawNominalType.java, so we can keep adding them even if the raw type is contained in a JSType.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=99001809