Improve our native handling of arguments array
We removed it from the externs several months ago, but didn't add it
to Scope.getVar(). This CL adds it there as well as across the compiler in
places that refer to Vars that may be Arguments.

Fixes #1125 on github
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=103199949