Stop claiming to support serialization of CompilerOptions and delete the OptionsSerializationTest.
Serializing options is not really supported (if you serialize options using one version of the compiler and try to deserialize using another, you'll likely get nonsense) and if we were to support something like that we'd use a specific JSON/proto/etc. format for serialization.

We were serializing and deserializing options in the OptionsFactory, but that was removed in cl/85441072.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=103333586