[NTI] Don't do possibly-inexistent-property warnings on objects with optional properties.
When the receiver has an optional property and accesses it, we currently give two warnings, one for possibly-inexistent property and one because the context usually expects a type that doesn't include undefined. No point in giving two, we should warn only for the 2nd thing, and the programmer can check for undefined if they want. The 1st warning is also bad because they can't cast to a sensible type to avoid it, only to ?.

For unions of objects, we still give the possibly-inexistent-property warning. For those, the type doesn't include undefined when accessing the property.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=112471095