Allow code to @export a list of tokens.
E.g.: /** @export */ ['a', 'b'];

This is useful for applications that use a reflection based templating language
to tell JSCompiler which tokens are ever used in a template expression. The
intention is to e.g. generate a list of tokens from Angular templates as part of
the compilation process of an app.