Move synthesized extern declarations back to window.
Now that NTI understands that window can be a namespace, it makes
sense to undo this, so that qualified names starting with window
can be used in type annotations.

This allows people to refer to the aliased name in JSDoc, e.g.
  /** @type {window.String} */
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=113178444