Fix class name not properly scoped to CLASS
We need to treat CLASS as a root for a block scope possibly
containing just the "class name" declaration.

See "classScope" in:
http://www.ecma-international.org/ecma-262/6.0/index.html#sec-runtime-semantics-classdefinitionevaluation

Add a few unit tests regarding class names and function names.

Note that our transpilation doesn't support referring to the class name
in a method body yet. Eg:

var Clazz = class Foo {
  bar() {
    return new Foo;
  }
};

is currently transpiled to:

/**@constructor @struct */var Clazz = function() {
};
Clazz.prototype.bar = function() {
  return new Foo;
};

which does not work at runtime. This will be fixed later.

Part of the fix to #1546.