[NTI] Improve handling of instanceof
If we know for sure that `x` is not of type `Foo`, we can infer that
`x instanceof Foo` is of `FALSE_TYPE`. On the other hand, if we know
for sure that `x` is of type `Foo` already, we can infer that
`x instanceof Foo` is of `TRUE_TYPE`.

Part of the fix to #1589.