Add warning for passing an Array to goog.object.forEach (and similar methods). goog.array.forEach should generally be used instead.
With polyfills enabled, this can cause runtime issues because the "values" property of arrays is enumerable, so it comes up when doing a for/in loop on an Array. (This is in violation of the spec, so we should also change the way the polyfills work. But on IE8 there is no way to set a non-enumerable property, so this check is still valuable.)
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=117260770