RefasterJS: Improve type matching of templates.
Instead of using loose matching, which can result in post-refactoring code that
does not type-check, use a strict type matching that matches only the exact
type specified in the template, or proper subtypes thereof.

To allow refactoring of code that does not quite match types as needed in the
after-template, as well as un-typed code, this CL changes RefasterJS to attempt
template matches strictly in the order in which the before-templates appear in
the template file.

This allows multiple before-templates of a given pattern to be specified with
successively more loose type matching, and appropriate after-templates.

See examples/navigational_xss_sinks.js for an extensive example.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=118178934