Add possibleMissingProperty to DiagnosticGroup
This new flag allows for disabling missing properties for unknown types.
This behaviour is desired when you have mixed typed and untyped objects.
In this case you still want the compiler to complain about missing
properties for known types but not for the untyped objects.

Example when this flag is disabled using
'--jscomp_error=missingProperty'
'--jscomp_off=possibleMissingProperty':

// Typed variable example
/** @constructor */
var A = function() {};
new A().foo; // This will emit an error about 'foo' undefined on 'A'.

// Untyped variable example
/** @type {?} */
var a = {};
a.foo; // No error is emitted.