Work around GWT parser's handling of empty annotations.
Currently the GWT compiler simply hardcodes the empty string for JVM properties, but this leads to the parser's "unknown annotations" list being parsed as a singleton list containing only the empty string, which is bogus.  The upshot is that a particular angular.js file has some JSDoc with a CSS example containing "@-webkit" which gets parsed as an empty annotation and breaks rhino's source position information, causing an unchecked exception and crashing the parser.  This change fixes it in two different places: (1) we no longer add the empty string as a valid annotation name, and (2) even if an empty-named annotation is found by the JsDocInfoParser, it is considered invalid.  It also adds better debugging to the exception thrown by SourcePosition, which previously omitted the line number (which turned out to be very important in tracking down the issue).
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=120186848