Refactor JSCompiler resource handling so that additional files can be added without having to add them specifically to the GWT-supersourced ResourceLoader.java.
The basic idea here is that we concatenate all the resources into a single file and the inject that single file as a resource.  The GWT resource loader then pulls apart the individual files into an ImmutableMap so that they can each be injected.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=120598689