For ES6 modules, when there is no CompilerInput corresponding to a given import statement, report a warning instead of an error.
This warning can easily be ignored when doing file-by-file transpilation, but will be visible to users when doing a full compile (in fact, it will be visible as an error, if they pass --strict).

For CommonJS modules, this is still an error, because a CommonJS module identifier can refer to a few different possible paths, and because I don't understand CommonJS as well so I want to minimize changes in that area.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=122294303