Tighten the dead property elimination deoptimzation
The previous implemention was a little too restrictive for most environments.
We now only trigger a deop if any of the following is true:

  a. A Object.defineProperty(...) where an aliased property name is being
  assigned to an aliased descriptor or the descriptor contains a getter/setter
  def.
  b. A Object.defineProperties where the last param is not an object literal.

To ensure safety, a property being defined with an aliased descriptor now just
blacklists that property name.

Finally, a new diagnostic error is introduced to report when there is a
pattern that triggers a critical deoptimization.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=124891113