Migrate polyfills to ES5.
Previously the polyfills were written in ES6 and transpiled down to ES5 (and concatenated) into es6_runtime.js during the compiler's build, which requires us to check in generated code into our repository.  In order to eliminate this pre-compilation step (and use the separate files directly), we need to manually bring everything down to ES5, since the libraries are injected part-way through transpilation, so if a library is injected later that requires an earlier pre-processing, we'd have a bad time.  It might be possible in the future to move back to ES6 but we'd need to take care to ensure the injected libraries get the full transpilation treatment, and it's not clear that there would be any benefit since the whole point of these polyfills is that the browser *doesn't* support ES6, so they're only used when language_out < ES6.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=124907074