[NTI] Reduce scope of CheckForInOverArray, rename to CheckArrayWithGoogObject, convert to use TypeI.
This deletes the lint against raw for-in-over-array (but keeps the lint
against passing an array to goog.object). This lint was written before Array implemented IArrayLike. With IArrayLike, we get a type error when using an array in a for/in.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=125382531