Allow dead property removal to run if the es6 runtime is present
The ES6 runtime for handling ES6 inheritance has a Object.defineProperty that
causes a deoptimization. However, this particular usage is not a concern since
if is actually copying a property that has a getter/setter, we would have
already seen it.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=125385891