Add support for 'syntactic' inference of /** @const */ types for .i.js files.
The idea of this is to allow .i.js generation to work better with type-checking
turned off, allowing faster builds.

This initial version only supports inferring the type of a literal RHS, but
it should be easy to extend this support to work on an RHS of a simple name
that has a declared type.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=127355199