Support uninlining the exports of a goog.module to create fewer aliases.
If a goog.module has exports of the form `exports = name` then we can create fewer aliases by replacing the definition of `name` throughout the module with the synthesized module export name, rather than synthesizing both a module-local and an export name.
So for example, instead of transpiling:

    goog.module('a.b.c');
    class Foo {}
    exports = Foo

to:

    class module$contents$a$b$c_Foo {}
    /** @const */ var module$exports$a$b$c = module$contents$a$b$c_Foo;

we can transpile to the even simpler:

    class module$exports$a$b$c {}

The intention is to use this to deprecate/remove the inlineAliases pass.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=128640756