Don't polyfill setPrototypeOf at all in IE<11 since it's not possible.
Previously we polyfilled it with a method that always threw or returned false.  But Babel seems to polyfill ES6 class-side inheritance by checking existence of Object.setPrototypeOf and using it if it's present.  So if we polyfill it then it will get used and break.  Instead, we can deal with this the same way we deal with Array.prototype polyfills in IE8 and simply refuse to polyfill them at runtime.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=128651909