Add support for multiple function definitions and specifically:
  something ? function() {} : function() {}

This allows, goog.string.trim and similar functions to be considered side-effect free and as a result most goog.userAgent methods to be considered side-effect free and removable.

This is mostly identical to the original except this version recurses properly to handle multiple levels of HOOK expressions and an additional unit test to cover the failing condition.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=130303575