Allow applying input source maps to the generated source map.
This allows users to easily map from generated source that's an
input to the compiler back into their original source files.

E.g. when compiling TypeScript, users would have:
    input.ts -(a)-> input.js -(b)-> compiled.js

Where (a) is the TypeScript compilation operation, and (b) is Closure
Compiler. With this option set, the source map generated in (a) is
applied to the operation (b), with the final source map mapping back
into `input.ts` from locations in `compiled.js`.

This also introduces an interface `SourceFileMapping` to avoid
increasing the coupling and responsibility of the Compiler object.