Improve ArrowType.hashCode and FunctionType.hashCode to make hash collisions less likely.
For ArrowType:
The previous version simply summed all the parameter hashcodes, but since + is commutative this means that function(a,b){} and function(b,a){} would have the same hashCode.

Also it included the 'returnTypeInferred' field in the hashCode though it was not part of equals().  This has been fixed.

For FunctionType:
All functions with the same ArrowType would get the same hashcode.  To resolve this we instead just hash more data.

Doing this revealed a bug in the equals() method for FunctionType where it was assymmetric and allowed constructors to be equal to ordinary functions in some circumstances.  This has been also been fixed.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=141193824