Skip unchanged functions in DeadAssignmentsElimination
Also, do the early bailout bookkeeping as part of the intial crawl
so that we don't need to do an extra crawl for it.

For large compiles, these can make the pass take about half the time as it did
previously.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=132491657