Fix enum type references in generated externs
When a generated extern referenced an enum type in a parameter or
record type, the primitive underlying the enum was being used instead
of the enum itself.  For example, this code:

  /** @enum {number} */
  var E = { A: 1, B: 2 };
  /**
   * @param {function(E)} takesEnum
   * @param {E} e
   */
  function f(takesEnum, e) {
    takesEnum(e);
  }

Would become this in the generated externs for the library:

  /** @enum {number} */
  var E = { A: 1, B: 2 };
  /**
   * @param {function(number)} takesEnum
   * @param {number} e
   * @return {undefined}
   */
  var f = function(takesEnum, e) {
  };

This would affect any application wanting to use the correct enum
types when referencing this library.

This change ensures that the enum name itself is used when referencing
the enum.  The primitive type under the enum is only used in the enum
declaration.

Related to #1399 on GitHub

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=133907989