Add a suggested fix for variables referenced before they're declared.
The fix is pretty naive: It just inserts a var statement. So:

  if (x) return true;
  ...
  var x;

becomes

  var x;
  if (x) return true;
  ...
  var x;

Ideally we'd also eliminate the later var statement, but since the ErrorToFixMapper doesn't have a NodeTraversal and Scope, that would be tricky to do. So, the fix we have here gets rid of the error but creates a different one ("Variable x is redeclared"). Then, you can run the fixer again, and it will delete the second var statement, using the SuggestedFix that was added in [] Your code will end up in good shape but only after running the fixer twice. This not ideal, but should work well enough to automate the fixes we need in d3: https://github.com/d3/d3/issues/2978

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=134700704