Make GeolocationPositionOptions a @record
As a @constructor, user code which calls Geolocation methods generates a warning. Since there is no user-callable constructor, this needs to be a @record in order to allow an Object with the given properties to be treated as valid.

Additionally, the three properties are supposed to be optional, so their types now allow undefined. (Should they allow null too?)
From the spec (https://dev.w3.org/geo/api/spec-source.html#position_options_interface):

> In ECMAScript, the enableHighAccuracy, timeout and maximumAge properties are all optional: when creating a PositionOptions object, the developer may specify any of these properties.

Also made position of type annotations consistent (on request).