Extracts aggressive alias inlining out of CollapseProperties, now it's own pass.
This is good both because it makes CollapseProperties more focused and because
it makes the aggressively inlined AST state available to other optimization
passes that want to be able to assume such an AST state as input.

In the past some passes (like J2clPropertyInliner) assumed that aliases would be
inlined before they saw the AST, and when this was not true the solution was to
tell users that it was their responsibility to mark type alias vars @const so
that the InlineAliases pass would handle them. But it is no longer the desire of
JSCompiler team to require this of users, which makes the change (being performed
in this CL) a necessity.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=136073701