When there are two insertions in the same place, make sure they are applied in the right order.
We don't actually need this now because the collections we're using in ApplySuggestedFixes keep items in insertion order. However, in an upcoming change, I'm switching from ImmutableSetMultimap to HashMultimap so that is no longer true, and it causes this to fail sometimes depending on iteration order of the HashSets.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=136752171