Add ^^^'s on the entire node associated with an error. I expect this will be helpful for users in some cases, e.g. if there is a conformance violation for an unknown type on "foo.bar.baz" it will be immediately obvious if it's "foo.bar" that is unknown or "foo.bar.baz"
Also (and this is my actual motivation for doing this :) it will help the compiler team to make sure errors are being reported on the exact node that is intended. This often matters when writing automatic fixes in ErrorToFixMapper.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=140510226