lint check: Enforce alias capitalization conventions in goog.module
If the right side is capitalized, indicating that it's a class, then the alias should be capitalized too. And if the right side is not capitalized then the alias should not be either.

In a later CL we may add further checks, such as ensuring that the alias name matches the last part of the fully-qualified name (except in cases like name collisions, or goog.string, where an exact match is inadvisable) but I want to start with this to make sure we don't overwhelm people by reporting their existing violations.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=150811935