Automated rollback of RescopeGlobalSymbols change
*** Reason for rollback ***

Breaks a few internal projects

*** Original change description ***

In RescopeGlobalSymbols, when a variable is defined both in externs and source, the externs definition wins.

To do this, we move the externs rewriting early, to happen at the same time as the rewriting of the other variables. This way, the traversal after RewriteScopeCallback needs no changes.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=151454658