Avoid unnecessary renaming in Normalize.
In MakeDeclaredNamesUnique, if we're traversing with the ES6 scope creator, don't traverse into blocks multiple times when looking for duplicate names.

With this change, running Normalize on a tree that's already normalized is a no-op, which is helpful because several tests run Normalize before and after the pass-under-test.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=152518777