Fixed a nonexistent URL and removed a redundant definition.
MemoryInfo is a nonstandard Blink-only interface. WebKit does not have it.
Removed the PerformanceMemory definition and changed the type of performance.memory, as it really is MemoryInfo, which already exists in webkit_dom.js.
Also clarified that it is nonstandard, Blink specific and exposed even without a flag.

Not adding @deprecated, because that will surely break Google projects, but it really should be.