Fixed a nonexistent URL and removed a redundant definition.
MemoryInfo is a nonstandard Blink-only interface. WebKit does not have it.
Removed the PerformanceMemory definition and changed the type of performance.memory, as it really is MemoryInfo, which already exists in webkit_dom.js.
Also clarified that it is nonstandard, Blink specific and exposed even without a...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=153621106