Fixes a "reported changes must have a scope" violation in RemoveUnusedVars.
RemoveUnusedVars was sometimes removing references inside of subtrees that
were already removed. Correct this in an elegant way is very difficult
because the different removals are tracked in different datastructures,
some of which are difficult to update for nodes in subtress and others of
which don't track the full set of node types.

I had a larger, less hacky and slower solution. But it didn't always work
since defFinder doesn't track GET_ELEM nodes and the other tracking
datastructures do.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=154899395