Fixes a "reported changes must have a scope" violation in OptimizeParameters.
It relies on a definition finder instance as an index within which to find
parameters to change/remove. But it was not updating the index to match
ongoing AST changes which sometimes lead to the index containing references
to already-removed nodes and this lead the duplicative attempts to remove
them and report the associated change.

The pass was making a cursory attempt to mitigate this by collecting all the
removed nodes into a list and then updating the index at the end, but it was
possible for the duplicate removal attempts to occur between the original
removal and before the late index update.

I changed the index updates to each occur immediately after each node
removal.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=154984241