[NTI] Simplify the unification algorithm; the new one is much easier to understand.
The old algorithm had the same semantics as the new one in the common cases, but was subtly wrong in some tricky cases. The new one is easier to describe declaratively, so easier to get right.

Fixed two bugs in the process: one with unification of enums, and one with unification of interfaces.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=156313351