Instanceof guards should not block code motion.
This change fixes a bug in CrossModuleCodeMotion that could potentially have a
big impact on how code is moved across modules.

When a reference looks like this
`x instanceof Foo`

CrossModuleCodeMotion changes it to this:
`'undefined' != typeof Foo && x instanceof Foo`

Once that is done, it is safe to ignore the reference when deciding where to
move Foo.

This change is to ensure that CrossModuleCodeMotion
ignores the added `'undefined' != typeof Foo` references when it is run
repeatedly.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=156437074