[NTI] Migrate TypedCodePrinter to NTI.
In performing this migration, a number of bugs on OTI's typed code printing surfaced (primarily around nullability), so I fixed them and added a few regression tests.

One omission was leaving out the `@this` annotation from functions.  Fixing this unfortunately introduces a few redundant `@this` annotations where they shouldn't be necessary (we currently do a bit of work to try to eliminate these, but it depends on type information in the AST which isn't always available - see a few test cases that have been changed; potentially it's not worth doing this at all and we could eliminate the extra logic and simply always emit `@this`).

A number of new TypeI methods were required to support this.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157625955