Remove most uses of Es6CompilerTestCase.
It is not simply a CompilerTestCase variant for ES6 code, but rather one that would run the same test twice, once with the compiler in ES6 mode and once in ES5 mode. This was helpful for passes that had logic like

if (compiler.getLanguageMode().isEs6OrHigher()) {
  // behave a certain way
} else {
  // behave some different way
}

to make sure both branches were covered. But, no passes have that sort of logic anymore.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157640378