Backoff of nested functions in the CheckMissingSuper pass.
It's acceptable to reference 'this' from within a function that is within an ES6 constructor before calling the super(). It will only be a runtime error if the function is called before the super() has returned. However, the JS cannot prove this, nor can it prove that 'this' is not bound to something entirely different. Therefore it should backoff in these cases.

This also uncovered a bug where the pass would accept the super() call from a nested class as the super() call for the enclosing class, this has also been fixed.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157739094