Add a featureSet method to PassFactory, to indicate which features the pass understands.
When running passes through PhaseOptimizer, a pass will be skipped if it's incompatible with the feature set currently present in the AST. For now, the default value of featureSet() is FeatureSet.latest() so this is a no-op. But it will soon change to ES5 so that users can pass --language_out=ES_2015, skip transpilation, and run ES2015-compatible passes.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157746761