Untangle the web between dependency crawling and module rewriting.
This results in being able to transpile all of the ES6 modules at once, rather
than in two separate phases within parseInputs().

This particular commit leaves the transpilation at the end of parseInputs(),
with the idea that it should be easy to move it out and into its own
compiler pass in a follow-up.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158540630