[NTI] Different fix for github issue #2520 (see []
Analyze the arguments in reverse in the backward direction, but add them to the builder correctly.

We analyze in reverse in several places in the backward direction, to preserve the order of side effects. Changing that only for loose calls wouldn't make sense.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=159029001