Improves j2clOptBundlePass perf by running on just changed scopes.
In [] j2clOptBundlePass accounts for 7.15% of execution time. 4.0% of that time is the PureFunctionIdentifier and the other 3.15% is J2CL pass visitors.

Though this CL does not migrate *all* J2CL pass visitors to the new scope limited traversal, it does migrate most (and documents what is necessary to migrate the rest) and reduces that 3.15% cost to 1.5%.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=159268542