Don't serialize cached source code in SourceFiles except for SourceFile.Generated.
In most cases the code can be easily refetched if needed in the second stage of a multistage compilation.
In the case of SourceFile.Generated, in order to recreate the cached source the Generator has to be run.
Classes that implement the Generator interface might have some state that needs serialization. Because these classes are extension points to jscompiler we opt for not requiring them to be serialized, and instead for these cases the cached source is serialized.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=159759402