Use the span of the function name for method defs in object literals and goog.defineClass.
For object literals:
Previously we used the span of the entire function for the STRING_KEY for a
function's name when it was a method def in an object literal. For example, in
'var foo = { f2() {}};', the span for f2 would be 'f2() {}'. With this change,
the span for the name will just be 'f2'.

For goog.defineClass:
We would use the root of the property as the truth for
source location information during the rewrite of the property name. When you
used foo: function()... syntax, this would work because the root is a STRING_KEY
for foo. However, if you used foo() {} syntax the root would still be the
MEMBER_FUNCTION_DEF, which has a span covering the whole function. In the
MEMBER_FUNCTION_DEF case, we actually want to use the first-first child (a NAME
node) for the source location information.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=160461876