Consolidate DisambiguateProperties to use TypeInvalidator.
There were a number of changes that needed to happen to TypeInvalidator for this to work, since disambiguation has rather significantly different criteria compared to ambiguation and inlining.  Additionally, the disambiguate pass has some warning rate-limiting logic shoe-horned into it, which needed to instead be added to InvalidatingTypes.  I hope to simplify some of this logic (and, in particular, the way nodes and errors are stored in TypeMismatch objects) in a future change.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=161612705