Produce octal value even when reporting a parse error due to strict mode.
For some tools that ignore parse errors it is useful to parse code assuming code is "strict", but retain octal numbers.  So, when we report a parse error for octal numbers go ahead a produce the octal value as if there were not a parse error.

This also allows us to switch the NodeUtilTest to run using the default language mode.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=163536166