[NTI] Migrate DeclaredGlobalExternsOnWindowTest.
There are two test cases that still fail in NTI.  It seems when 'window' is not explicitly declared in externs, then NTI does not recognize types declared on it.  It's actually more surprising that OTI *does* recognize these types, since the pass seems like it should be a no-op in this case.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=165783100