Removes clinits more aggressively in J2clClinitPrunerPass.
When there is a deep class hierarchy, even the clinits are empty, we can only remove one clinit at each pass.
This makes the number of optimization loops to be at least same as the level of depth of class hierarchy and each loop is quite expensive since a lot of passes not looking at only changed code.

This patch updates clinit pruning pass to be more aggressive and remove clinits/references to a fixed-point in single pass. As a result we can remove  whole empty chain in the class hierarchy in single pass instead of relying on other passess to cleanup to be able to do more removals.

This CL drops number of runs in gmail by 11 and reduces the stage 2 total time by 10%

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=166369426