Prevent calls to getLastCompiler() from TypeICompilerTestCase.
When TypeInferenceMode is BOTH, this call is practically guaranteed to be an error.  The entire purpose is to do some additional check on the AST or some other aspect of the compile, but in BOTH mode, there are two different compilers created, and only the second one (NTI) is accessible afterwards.

The fix is to add a new Postcondition TestPart to CompilerTestCase, which contains a callback that accepts the Compiler and is called at the end of CompilerTestCase#testInternal, so that all test cases can be verified separately.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=166652077