Call assertAbout() instead of directly calling Subject constructors. Direct calls require access to a FailureStrategy instance, and we're hiding those instances.
Additionally, in this particular CL, call check().about(...) instead of directly calling a Subject constructor with THROW_ASSERTION_ERROR.
Avoiding the Subject constructor is good for the reasons described above. But switching specifically to check().about(...) rather than assertAbout(...) also means that, if someone uses the Subject with a different failure strategy, we'll correctly propagate that strategy to derived subjects, rather than switch back to THROW_ASSERTION_ERROR.

I know that this CL adds a lot of boilerplate. If it's any consolation:
- We plan for check() to automatically add context in the future, so you'll get some value from it.
- We plan to convert SubjectFactory to an interface, so we'll come back and convert these all to "public static SubjectFactory<FooSubject> foos() { return FooSubject::new; }."

Sorry for the mess in the meantime.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=168012115