Always transpile modules, even if the output language requested was ES6 or higher.
Since the compiler outputs a single file, it doesn't make sense to leave the import/export statements alone, unless the input to the compiler was just a single file, which is very rare.

This allows for projects that are using ES6 modules to pass --language_out=ES_2015 or higher.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=172155957