Automated g4 rollback of changelist 172155957.
*** Reason for rollback ***

Need to roll this back to unblock rollback of #2641

*** Original change description ***

Always transpile modules, even if the output language requested was ES6 or higher.

Since the compiler outputs a single file, it doesn't make sense to leave the import/export statements alone, unless the input to the compiler was just a single file, which is very rare.

This allows for projects that are using ES6 modules to pass --language_out=ES_2015 or higher.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=172167855