[NTI] Add an "unresolved type" to the type system.
This type is useful for per-library type checking. The semantics is that this type is neither a subtype nor a supertype of any other type. In effect, if you try to use it, you get a warning.

To avoid flowing this type around and having to change all type operations to recognize it, if an expression evaluates to this type, we warn and return ? instead.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=172934055