Automated g4 rollback of changelist 175230172.
*** Reason for rollback ***

Breaks a bunch of stuff: []

*** Original change description ***

Refactor RemoveUnusedVars to allow for logic to remove properties.

Along the way a few bugs were discovered and fixed with test cases added for
them.

The logic that actually removes unused properties is disabled here,
but present.  Tests for it will be added in a followup CL.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=175292210