Simplify Es6ExternsCheck to not special-case BLOCK_SCOPED_FUNCTION_DECLARATION.
As far as I can tell, this is completely equivalent.  The Feature is declared as ES3, but it looks like this special casing logic was written assuming it was declared ES6.  Since it's not declared as ES6, removing the Feature before re-checking that the featureSet is a subset of ES5 is irrelevant.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=178416736