Fix the way we use BlackHoleErrorManager
I discovered that setting this on the CompilerOptions won't actually silence warnings. To do that you need to pass a BlackHoleErrorManager to the constructor of *Compiler* or to its setErrorManager() method.

(Eventually, I would like to unify the ErrorHandler which is a field in CompilerOptions and the ErrorManager which is a field in Compiler, because the current state of things is definitely confusing)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=178986028