Represent the function inlining level as a Reach
This is the API that we are encouraging, so we may as well represent our
options this way internally as well.

I propose we leave the setter setInlineFunctions(boolean) until after the next
external release before we remove it, since it is quite old and can still be
implemented in a reasonable way.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=179590828