Remove the NodeUtil.isAssignmentTarget
It is of limited scope (only used in a single pass) and confusingly similar to
NodeUtil.isLValue.  It looks like all the existing uses can just be converted
to use isLValue, so we convert them as well in this change.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=179949354