Implement TypedScope.makeArgumentsVar().
Previously TypedScope,getVar("arguments") would return a var from the outer scope (typically the es3.js externs file), but this means that all functions share the same arguments Var, which is incorrect.  It's possible that this does not cause any problems, since it only affects type checking, but it's also possible that it could lead to subtle bugs if it's not fixed.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=181700817