Create $jscomp.assign() for use within the compiler
The method is equivalent to Object.assign(), but guaranteed to
be available even if polyfills are not included in the
compiler output.

This is desired for use in Es6RewriteBlockScopedDeclaration
to avoid use of explicit object literals to update loop scopes.

Related to https://github.com/google/closure-compiler/issues/2779

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=183462508