For @externs and @typeSummary, print one per input rather than one per output
For bundling this doesn't always make sense, but there isn't a reasonable
way to bundle externs and non-externs into the same file anyway, and this way
is more in line with how it would work to bundle a single file at a time.

This also makes .i.js generation work the same way when given a single file
as it does when given multiple files.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=185202801