Fix for InlineFunctions issue with lets in for loops.
We never have to consider the case "for (var i = inlinableFn(); ..." because Normalize will have already removed the var from the loop. But it doesn't do this for "for (let i = inlinableFn(); ..." so we need to be careful not to treat that case as a VAR_DECL_SIMPLE_ASSIGNMENT.

It still does end up getting inlined, but only after being decomposed by the ExpressionDecomposer.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=186544239