Roll forward: Update DepsGenerator to support goog.require's with es6 modules:
- goog.require(path) is now allowed in goog.modules for ES6 modules.
- ES6 modules no longer provide any symbols in a deps file.
- Requires for ES6 modules in a deps file will be their relative path to Closure rather than a munged symbol.

NEW:

- Emit a warning if a file is any combination of goog.provide, goog.module, ES6 module. TODO for this to become an error.
- Filter out munged provides when writing ES6 modules to a deps file, leaving in place any other provided symbols. Before ES6 modules just always had no provides, no matter what.

The above means that an ES6 module with a goog.provide is technically warning and should work, for now. Once we have more official mechanisms for interop this will be promoted to an error. *Do not rely on the behavior because it will eventually be an error!*

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=187353966