Report an undefined variable error for "() => y" in VarCheck
Previously VarCheck was allowing anything that was the child of a function expression to be an undeclared name, not just the function name. Now it correctly reports an error for "() => y" and adds the undeclared name "y" to the externs.

Fixes https://github.com/google/closure-compiler/issues/2843

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=188941589