Streamline special handling in Var.Arguments to also handle 'this' and 'super'.
Adds an AbstractScope.ImplicitVar enum that includes the three special vars that are implicitly created in different scopes.  The enum provides a method to know whether any given scope creates the var, and is integrated into the various Scope.getVar() methods to do the right thing.

This is a step toward supporting ES6 block scopes in LinkedFlowScope, since we need to be more careful exactly which scope to declare qualified names on (there's no longer just the function scope, but the various block scopes in between).  The general approach is to look for the root of the qualified name and this streamlines the case where the root is 'this' or 'super' to not need any special handling.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=189380979