Rephrase error message for JSC_MISSING_RETURN_JSDOC
Currently it's a bit unclear when an inline return annotation is accepted by the compiler, and separately, when it's encouraged by the style guide. The purpose of the lint check is to ask the user to indicate the return type in some way, rather than to specify exactly how they should do that.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=189961196