Update parts of TransformStreams externs to match spec.
Intentionally leaving the definition of `TransformStream` as a @record for now, even though it should be constructor according to the spec. This is because a bunch of code already relies on this @record definition.

This is a second attempt at this change

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=191254581