Rename recordInvalidations to withInvalidationMap.
The name recordInvalidations implies that the class should pass a non-empty map of invalidations which will be recorded. In fact, when this is called, the caller passes a mutable empty map, which the Builder later uses to record invalidations.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=191763716