Add parentheses around "{" at the start of an arrow function body
We already had logic to add parentheses when the object literal was the only thing being returned, but not when the object literal or object pattern was more deeply nested in an expression.

To do this I added two states to the CodeGenerator.Context enum. One is for when we are at the start of an arrow function body and the other is for the case where we are both in an arrow function body and inside a for loop init clause.

(Deciding to add parentheses by recursing to see the leftmost child is an object lit caused some unnecessary parentheses: e.g. in cases where something like a cast also added parentheses before the "{", I was adding parentheses twice)

Fixes https://github.com/google/closure-compiler/issues/2883

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=194014420