Don't update the TypedScope for empty let declarations with no declared type
Specifically: when we see a let declaration with no type and no value, e.g. "let foo;" in TypeInference, we do not change its type on the TypedScope to be undefined. This allows some sloppier typing of lets that are declared, later assigned, then later used in callbacks.

This is to be compatible with how we currently handle VAR declarations and unblock natively typechecking let and const. It's not the desired end-state, though.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=194329783