Automated g4 rollback of changelist 195478973.
*** Reason for rollback ***

When the API changed from Path to URI calling code used path.toURI to convert. However this was a behavioral change that ends up including the file:// scheme. In the roll forward the equivalent code is actually new URI(path.toString()).

*** Original change description ***

Allow the transpiler to take URIs so that URLs may be used. Useful for servers that output sourcemaps to ensure the sourcemap points to a server file and not one on disc.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=195713404