Automated g4 rollback of changelist 195893567.
*** Reason for rollback ***

Clean rollforward now that clutz has been patched to handle this behavior.

*** Original change description ***

Automated g4 rollback of changelist 195872147.

*** Reason for rollback ***

Breaks some typescript builds

*** Original change description ***

Declare externs on the global `this` when window is not present.

This fix an issue where externs weren't assumed to be propeties on any object
when `window` is absent from the externs. This could cause the compiler to
improperly rename properties which are actually references to externs.

***

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=196035069