Partial rollback of "Make ES6 module bundling work with different module resolution schemes."
The issue with the change is that registered paths are no longer escaped,
but import paths are.  So these paths don't line up.
In order to roll this forward it requires a semi large reworking of the module
loader to be configured to not escape paths.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=197468860