Remove unnecessary @CheckReturnValue from NodeUtil methods
The point of this annotation is to make it clear that it's not mutating the receiver or argument.  But these methods are static and only accept Node.  The method names in no way suggest that the nodes would be mutated to instead point to an enclosing node.  So these annotations instead end up as additional boilerplate and/or useless confusion/complexity.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=197576222