Remove special member collection in TypedScopeCreator.LocalScopeBuilder
As noted in the TODO, this was papering over an issue where we skipped declaring properties on instance objects.  The correct fix is to also declare the property if the owner is THIS.

This clears the way for unifying LocalScopeBuilder and GlobalScopeBuilder into a single NormalScopeBuilder.  We can then pull out FunctionScopeBuilder and ClassScopeBuilder with significantly less logic in them.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=198132934