Add Promise and Thenable as native types
This is in preparation for natively typechecking async functions, as we want to access the native promise type for typechecking async function returns.

We also need to add Thenable because it's used in the parameters for the Promise constructor.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=199539818