Export the gjd method of JsfileParser in the same JS module as the GwtRunner.
This are slightly awkward since GwtRunner already is using default exports. So the module is the compile function that has the gjd function as a property. To make things slightly less confusing I've also added the compile function as a property of itself, so that the module may be treated either as the compile function or namespace with the compile and gjd functions.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=199863733