Avoid passing around names with non-identifier chars in them
This fixes a bug where we were looking up names like "<" in the type registry, which throws an IAE because it thinks it was given a templated name, which is not allowed.  This change fixes it conservatively by still returning names for quoted properties, but only if they are valid JS identifiers.  We might want to go further by returning null for *all* quoted properties, but this is a bigger change and might break other things.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=200105915