Roll forward of "Export the gjd method of JsfileParser in the same JS module as the GwtRunner."
This are slightly awkward since GwtRunner already is using default exports. So the module is the compile function that has the gjd function as a property. To make things slightly less confusing I've also added the compile function as a property of itself, so that the module may be treated either as the compile function or namespace with the compile and gjd functions.

NEW: Don't have any static initialization code in GwtRunner. I have no idea how my change broke this, but it did. After my change the static intializer ran earlier than before, and $wnd.Object was not available at that time ($wnd={Error:{}} to begin with and is set to "this" later). I attempted to make the GwtRunner variable in GwtRunnerMain non-static (even move it to the function), but that did not work.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=200462544