Update `hashCode()` on `TemplatizedType` to be equal to that of the corresponding `IntanceObjectType` when all template parameter values are the unknown type. (i.e. `hashCode(Foo) == hashCode(Foo<?>)`)
This is being done because currently such pairs of types (a raw template type and it's specialization on unknown) are considered identity equal (`equals()`). Therefore, the old behaviour violated the contract of `hashCode()`.

In a better world we would rather have `Foo !== Foo<?>`, but that change has larger implications. This CL represents a quick fix to bring things to a less-bad state.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=200632664